Drop remains of Minischedule.Load into Data
